sync with upstream

+++ enterprise/b/containers_btree.go
@@ -23,8 +23,8 @@ import (
 	"github.com/pilosa/pilosa/roaring"
 )
 
-func cmp(a, b uint64) int {
-	return int(a - b)
+func cmp(a, b uint64) int64 {
+	return int64(a - b)
 }
 
 type bTreeContainers struct {
@@ -88,13 +88,13 @@ func (u updater) update(oldV *roaring.Container, exists bool) (*roaring.Containe
 // this struct is added to prevent the closure locals from being escaped out to the heap
 type updater struct {
 	key           uint64
-	n             int
+	n             int32
 	containerType byte
 	mapped        bool
 }
 
 func (btc *bTreeContainers) PutContainerValues(key uint64, containerType byte, n int, mapped bool) {
-	a := updater{key, n, containerType, mapped}
+	a := updater{key, int32(n), containerType, mapped}
 	btc.tree.Put(key, a.update)
 }
